Fronts.
Fronts provide a common interface to a display front-end. Implementation-specific structures (such as SDL textures) will not be available to the rest of the code. This will help encapsulate code better and make a possible change to different display front-ends easier (e.g. interfacing directly with the display on a Raspberry Pi).