Implemented basic, incomplete mapper framework.
Mappers are located in mappers.c/.h and can interfere with CPU/PPU reads and
writes as is required. Currently no mapper but mapper 0 is implemented, as we
want to get one game working first. Games using mapper 0 are Donkey Kong and
Donkey Kong 3, among others. After this, mapper 4 in order to get Super Mario
Bros 3 to work. The system currently doesn't check whether a mapper is valid, so
unless the mapper number is 0, the program will segfault.

4-screen mirroring is currently not supported, since this is mapper-dependent.
Mappers are able to write to pointers that are used to look up data to make
common things such as bank-switching easy. If the need exists to 'listen' to a
memory location/register, then the mapper may interfere with loading and
storing. Currently the mapper has the lowest priority, so if the requested
memory location falls within one that is expressed through the pointer struct,
then the pointer struct will be used.

rom.c/.h contains code to load a ROM into a mapper struct, which also manages
the memory. Now we will have to merge the CPU and rest of the project to use the
mapper. In the case of the CPU, this will likely involve calling the mapper in
read8/write8 rather than accessing the array.

The mapper also needs to interact with the PPU registers, however Aurel is
handling the storage of this himself and will provide a method to call to copy
data through the mapper.