Set up basic functions for APU to talk to SDL
SDL's callback function will be called independently of the APU clock.
The caller will store the last value that was computed by the APU, and
the callback will fill the buffer with the current APU value. This way
the sound will change at the rate of the APU, but the playback will be
independent of the APU. Later we will likely be able to reduce the
sampling rate, which will make things more efficent.