Call apu_cycle every cycle.
The APU will then decide how to update the last_buff_val, which is what
determines the frequency currently being played. Some wave functions are
clocked at different rates to the CPU, but the APU will handle this
itself. Currently apu_cycle returns the new output value, but Andrei is
currently changing this so that it simply assigns last_buff_val itself
and the return type will soon change to void.