Notify APU of memory writes to register area.
The APU is expected to not write during the handling of this operation.
The APU should perform all memory writes during the cycle. The register
area isn't entirely owned by the APU, so the apu_mem_write function will
also be triggered for registers that the APU is not interested in, will
have to ignore these as is appropriate.