Added toggleable IS_PI flag.
If IS_PI is not set, then functions in the controller.c/.h files are
stubs. We can implement code that interacts with these without
experiencing any ill effects. I've also added a function that returns
whether a controller is available (i.e. whether the IS_PI flag is set)
at runtime.

To set the flag in CMake, provide -DIS_PI:BOOL=ON when running CMake.
The default value is OFF, for convenience.