Work in progress envelope + pulse channels.
As fas as I understand, the envelope controls the overall volume of a
channel. There's a constant volume option which basically negates
everything that it does, but oh well. The envelope is implemented now,
so I just have to see how I can make it fit in with the various
channels.