PPU properly displaying tiles now.
Confirmed with DK, DK3, and Bomberman (although Bomberman crashes the CPU immediately after the first frame). The background rendering was refactored a couple of times (currently it is in quite a readable form) before the actual problem was discovered - the bitfield structs used in the PPU were not "packed". This probably lead to issues when assigning raw values to said structs, especially when multiple bitfields were combined in the same union.

Sprite rendering is disabled for now, it is the next priority.

The UI now includes MMC CPU and MMC PPU buttons to display the mapped memory of the CPU and the PPU respectively. The PPU button now toggles the palette display.