WIP APU pulse 1/2
Implemented envelope & sweep unit, refactoring a bunch to remove code
duplication. Will implement pulse first and refactor again, as there are
several timers and structs which are redundant. I'm also currently not
happy with the way that registers are queried, but I'm implementing the
APU as I read the nesdev pages.