Keyboard controls now enabled on the SDL front.
The bindings are currently hardcoded and they are WSAD and arrow keys for the D-Pad, J for the B button, K for the A button. The IO tab now shows a sprite with the currently pressed buttons on both controllers, although only the first controller is controllable with the keyboard. The NES controllers via GPIO should still work, although I could not test this without the RPi.