Completed frame counter.
Only the mixer left to implement now, and then maybe it'll work. Also
I'm not implementing the DMC for now, since this operates differently to
the rest and isn't strictly necessary. I'll implement DMC once the rest
is confirmed working.

The APU has a lot of quirks and edge cases. I hope I've covered most of
them. In the frame counter, the interrupt may be a cycle off. Hopefully
this doesn't matter too much.

APU is now initialised properly, sweep channels hold a pointer to the
timer period since the period is 'sent' from the timer to the sweep
channel. If the sweep channels do end up having their own period (i.e.
the sweep modifies only its own period, but passes this on as the final
period) then I will have to change the sweep to hold its own period.
Also, the status register is actually not stored in 'memory', since its
flags are listened to and used a lot. It is stored internally as a
bitfield.